+++ CHANGELOG.md
@@ -3,8 +3,11 @@ All notable changes to **APCu Manager** are documented in this *changelog*.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and **APCu Manager** adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased - will be 1.0.1]
+## [Unreleased - will be 1.1.0]
+### Added
+- Compatibility with PerfOps.One suite.
 ### Changed
+- New menus (in the left admin bar) for accessing features: "PerfOps Analytics", "PerfOps Tools" and "PerfOps Settings".
 - Analysis delta time has been increased to avoid holes in stats when cron is not fully reliable.
 ### Fixed
 - With some plugins, box tooltips may be misplaced (css collision).